pyproject.toml
[project]
name = "procedural-human"
version = "1.0.0"
description = "Procedural Human Generator for Blender - Create realistic human models using Geometry Nodes"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Procedural Human Team", email = "info@proceduralhuman.dev" },
]
keywords = [
    "blender",
    "3d",
    "procedural",
    "human",
    "geometry-nodes",
    "modeling",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics :: 3D Modeling",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    # Development tools
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
    "fake-bpy-module>=20251003",
    "tree-sitter>=0.20.0",
    "tree-sitter-python>=0.20.0",
    "filemover @ git+https://github.com/plex1/filemover",
    
    # SAM3 core dependencies
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.35.0",
    "timm>=1.0.17",
    "numpy>=1.26,<2",
    "tqdm",
    "ftfy==6.1.1",
    "regex",
    "iopath>=0.1.10",
    "huggingface_hub",
    
    # Segmentation workflow dependencies
    "opencv-python>=4.8.0",
    "scikit-image>=0.21.0",
    "pillow>=10.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
# Note: PyTorch with CUDA requires manual installation matching user's CUDA version
# Example for CUDA 12.1: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
cuda = []

[project.urls]
Homepage = "https://github.com/procedural-human/blender-addon"
Documentation = "https://procedural-human.readthedocs.io/"
Repository = "https://github.com/procedural-human/blender-addon.git"
"Bug Tracker" = "https://github.com/procedural-human/blender-addon/issues"

# Build system (optional - use package.py for Blender addon distribution)
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.hatch.metadata]
allow-direct-references= true
